+++ src/ScottPlot5/ScottPlot5/Plottables/PolarAxis.cs
@@ -32,6 +32,11 @@ public class PolarAxis : IPlottable, IManagesAxisLimits
     public double PaddingFraction { get; set; } = 1.1;
 
     /// <summary>
+    /// Rotates the axis from its default position (where 0 points right)
+    /// </summary>
+    public double RotationDegrees { get; set; } = 0;
+
+    /// <summary>
     /// Enable this to modify the axis limits at render time to achieve "square axes"
     /// where the units/px values are equal for horizontal and vertical axes, allowing
     /// circles to always appear as circles instead of ellipses.
@@ -125,6 +130,7 @@ public class PolarAxis : IPlottable, IManagesAxisLimits
     /// </summary>
     public Coordinates GetCoordinates(double radius, double degrees)
     {
+        degrees -= RotationDegrees;
         double x = radius * Math.Cos(degrees * Math.PI / 180);
         double y = radius * Math.Sin(degrees * Math.PI / 180);
         return new Coordinates(x, y);
@@ -180,9 +186,10 @@ public class PolarAxis : IPlottable, IManagesAxisLimits
         using SKAutoCanvasRestore _ = new(rp.Canvas);
         Pixel origin = Axes.GetPixel(Coordinates.Origin);
         rp.Canvas.Translate(origin.X, origin.Y);
+        rp.Canvas.RotateDegrees((float)RotationDegrees);
 
         using SKPaint paint = new();
-        Spokes.ForEach(x => x.Render(rp, Axes, paint, PaddingFraction));
+        Spokes.ForEach(x => x.Render(rp, Axes, paint, PaddingFraction, RotationDegrees));
         Circles.ForEach(x => x.Render(rp, Axes, paint));
     }
 }